Changed table "Composite" to "CmpSummary" (ditto _East) so that all composite tables have the same nameing convention (that is, they start with Cmp). Sorted the list of available tables in the pick list.
Added a CmpMetaData table that lists the runs that were selected for the composites. 

Removed the use of the ":memory:" database I called "m." and I use the automatic memory database
that sqlite3 builds called "temp". I had to stop using dbListTables and replaced it with a query 
on sqlite_master.


git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/rFVS@2103 90b89b48-9230-0de4-5065-812dd12741b8