If a user previously uploaded a database where the “Stand_CN” and/or “StandPlot_CN” columns were not present in the database tables, runs would not execute. This has been addressed and it is no longer a hard requirement.
The "Stand Table" dropdown has been renamed to "Inventory Data Tables"

The "FVS Output" button have been renamed to "FVS Main Output File"

Progress bars have been added to the “Make New Project” action button, and to the loading of the “Help” tab.

3 bug fixes were addressed:

•	SPGroup disappearing from dropdowns after cutting/pasting
•	Component collection list not showing up after restarting
•	“Run Title” disappearing after duplicating a run

A popup warning now shows up if a user selects, and/or tries to "Save and Run", a run created with a database that is no longer the active uploaded database. After clicking "OK", the user is automatically re-directed to the "Upload Data" tab.






git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/rFVS@2703 90b89b48-9230-0de4-5065-812dd12741b8