Added code to have the "Run" tab become emboldened with two asterisks on either side so as to make it clear to the user that the run will need to be rerun in order for changes made to reflect in the output. This triggers whenever a stand has been added, the time scale has changed, a component has been added, an output has been selected/deselected, and whenever a simulation content element has been cut and/or pasted.
Using logic similar to Suppose, warnings were added in the instance that the the user enters values in the "Time" tab that are out or range.

Added code to grey out the tabs in Step 2 of the “Import Data" tab while the new database is being uploaded. Also changed the messages to be popup warnings. The warnings for installing/merging the new/training /blank databases have updateTabsetPanel calls to redirect to a new run (if "Yes" is selected) or to the edit and view tab in the case of the blank database (if "Yes" is clicked as well).


git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/rFVS@2724 90b89b48-9230-0de4-5065-812dd12741b8