rPkgsDev/fvsOL: Several bugs were fixed:
The code that finds the available FVS dlls was made stronger so that
files that are not dlls can be in the FVSbin directory without causing a problem.

When "New" run was pressed, sometimes the title was not being correctly
shown in the list of runs (very odd gibberish was shown).

The code in Duplicate run was causing the MgmtID to not be updated in the 
duplicate.

When a different run was selected, while a component was being built or
edited, the component was not being cleared from the display.

There were situations where the list of "categories" of management, modifiers,
or keywords were not being updated when a new run was created or one duplicted.


git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/rPkgsDev/fvsOL@3501 90b89b48-9230-0de4-5065-812dd12741b8