fvsOLdev: Moved the check on the presence of FVS_GroupAddFilesAndKeywords in uploaded data to where it is critical. The table is still added if neede but not changed if it is already present in the existing data and more data is added.
Fixed an issue where DBH is used in FVS_TreeInit in imported data and DIAMETER is already
in the table or "also" in the table. The not null DBH's are used to set the DIAMETER values
where the DIAMETER values are null.

When imported data has blank Stand_CN values they are set equal to the Stand_ID values
so that queries FVS uses to fetch data will work regardless of which field is used as a key.

If the "Groups" field is null, it is set to All_Stands in the FVS_StandInit table and set
to "All_Plots" in the FVS_PlotInit table.

The FVS_Data.db.default database is updated.



git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/Rcode@3940 90b89b48-9230-0de4-5065-812dd12741b8